Initial GUI version of the IDA Plugin
EVAL: Add support for C/C++ numeric values ending with 'l' too.
BASE: Ignore C/C++ suppossedly numbers that aren't in a numeric format because we have partial source code and the value was assigned using a (non resolved) macro.
CLANG: Improved support for handling partial source code.
GUI: Added a dialog for selecting the exported source code database, a chooser for showing the matches as well as support for diffing pseudo-codes and source codes.
MAIN: Added support for getting a similarity ratio for matches as well as to compare 2 functions.